Separate out convnet
